add highlighting for integers
Just handles integers (i.e. sequences of ASCII digits, "[0-9]+" as a
regex) for now, no decimals. So "12.34" gets interpreted as the integer
12 followed by a string "." followed by the integer 34.

Is also overly 'greedy', in that if we have
"Foo {foo1 = 'x', foo2 = 'y'}" say, the '1' in "foo1" gets treated as
an integer. As this is Haskell, we should probably respect "foo1" as an
identifier and leave the '1' alone.