fix issue with 'greediness' in parsing integers
As an example, consider the Haskell expression
"Foo {foo1 = 123, foo2 = 456}". Previously, we would treat the '1' in
"foo1" and the '2' in "foo2" as integers and highlight them
accordingly. With this commit, we instead consider strings like "foo1"
(that is, Haskell-style identifiers) as a plain string and leave the
'1' alone. (To be clear, we still highlight the 123 and 456 as integers
though.)