Preserve whitespace characteristics of input
Rather than stripping all whitespace after parsing, then adding it back between every expression, we now mostly keep whatever whitespace existed in the input.

This partially reverts 20d8a2d. With the other changes in that commit, we had assumed that this sledgehammer approach to whitespace was virtually necessary. This commit shows that, with a bit more care, that needn't be the case. We add some tests for inputs on which we had regressed to unpleasantly whitespace-heavy output.