Move away from build-type Custom
cabal-doctest is no longer unsupported or needed.

The author of the `doctest` package recommends[^1] to run doctests with

  cabal repl --with-ghc=doctest

This change removes the custom-setup and adjusts the test-suite to use
the recommended approach above. The package `doctest` is added to the
project with `extra-packages` to make sure its dependencies are always
in scope (a small variation to the approach suggested by the
`doctest` author).

[^1]: https://github.com/sol/doctest#running-doctest-for-a-cabal-package